Re-align use of signed/unsigned/long, use off_t and re-align format parameters.

+++ tools/apriori/apriori.h
@@ -4,7 +4,7 @@
 void apriori(char **execs, int num_execs,
              char *output,
              void (*set_next_link_address)(const char *name, off_t fsize),
-             int (*get_next_link_address)(const char *name),
+             off_t (*get_next_link_address)(const char *name),
              int locals_only,
              int dry_run,
              char **lib_lookup_dirs, int num_lib_lookup_dirs,
